zdded more testing and click handling for tasks

+++ composeApp/src/commonMain/kotlin/com/braiso_22/bichota_timer/tasks/domain/usecases/GetWorkedHoursOfTasks.kt
@@ -5,18 +5,18 @@ import com.braiso_22.bichota_timer.tasks.domain.entities.Segment
 import com.braiso_22.bichota_timer.tasks.domain.entities.Task
 
 class GetWorkedHoursOfTasks {
-    operator fun invoke(tasks: List<Task>): Float = tasks.getTotalHours()
+    operator fun invoke(tasks: List<Task>): Double = tasks.getTotalHours()
 
-    private fun List<Task>.getTotalHours(): Float {
+    private fun List<Task>.getTotalHours(): Double {
         val sumOfMinutes = this.sumOf { task -> task.totalMinutes() }
-        return sumOfMinutes.toFloat() / 60
+        return sumOfMinutes / 60
     }
 
-    private fun Task.totalMinutes(): Int = this.executions.sumOf { execution ->
+    private fun Task.totalMinutes(): Double = this.executions.sumOf { execution ->
         execution.totalMinutes()
     }
 
-    private fun Execution.totalMinutes(): Int = this.segments.sumOf { segment ->
+    private fun Execution.totalMinutes(): Double = this.segments.sumOf { segment ->
         segment.totalMinutes()
     }
 
@@ -24,7 +24,7 @@ class GetWorkedHoursOfTasks {
     The duration is 0 if the end is null because
     we don't want to have it in account until its finished
      */
-    private fun Segment.totalMinutes(): Int = this.end?.let { end ->
-        (end.toSecondOfDay() - this.start.toSecondOfDay()) / 60
-    } ?: 0
+    private fun Segment.totalMinutes(): Double = this.end?.let { end ->
+        (end.toSecondOfDay() - this.start.toSecondOfDay()).toDouble() / 60
+    } ?: 0.0
 }
